refactor: split CallButton into CallEndLeaveButton
- logic is only used in a single place and only when isInCall
