+++ src/components/CallView/BottomBar.vue
@@ -25,9 +25,9 @@ import IconSubtitles from 'vue-material-design-icons/Subtitles.vue'
 import IconSubtitlesOutline from 'vue-material-design-icons/SubtitlesOutline.vue'
 import IconViewGalleryOutline from 'vue-material-design-icons/ViewGalleryOutline.vue'
 import IconViewGridOutline from 'vue-material-design-icons/ViewGridOutline.vue'
-import CallButton from '../TopBar/CallButton.vue'
 import ReactionMenu from '../TopBar/ReactionMenu.vue'
 import TopBarMediaControls from '../TopBar/TopBarMediaControls.vue'
+import CallEndLeaveButton from './CallEndLeaveButton.vue'
 import {
 	toggleFullscreen,
 	useDocumentFullscreen,
@@ -644,7 +644,7 @@ function changeView() {
 				</NcActionButton>
 			</NcActions>
 
-			<CallButton
+			<CallEndLeaveButton
 				v-show="!(isVoiceRoom && isGuestActor)"
 				class="call-button"
 				:hideText="isSidebar || isMobile"
