+++ src/components/TopBar/CallButton.vue
@@ -3,100 +3,6 @@
   - SPDX-License-Identifier: AGPL-3.0-or-later
 -->
 
-<template>
-	<NcButton
-		v-if="showStartCallButton"
-		:title="startCallTitle"
-		:aria-label="startCallLabel"
-		:disabled="startCallButtonDisabled || loading || isJoiningCall"
-		class="join-call"
-		:variant="hasCall ? 'success' : 'primary'"
-		@click="handleClick">
-		<template #icon>
-			<NcLoadingIcon v-if="isJoiningCall || loading" :size="20" />
-			<IconPhoneDialOutline v-else-if="isPhoneRoom" :size="20" />
-			<IconPhoneOutline v-else-if="silentCall" :size="20" />
-			<IconPhone v-else :size="20" />
-		</template>
-		<template v-if="showButtonText" #default>
-			{{ startCallLabel }}
-		</template>
-	</NcButton>
-
-	<NcButton
-		v-else-if="showLeaveCallButton && canEndForAll && isPhoneRoom"
-		:aria-label="endCallLabel"
-		class="leave-call"
-		variant="error"
-		:disabled="loading"
-		@click="leaveCall(true)">
-		<template #icon>
-			<NcLoadingIcon v-if="loading" :size="20" />
-			<IconPhoneHangupOutline v-else :size="20" />
-		</template>
-		<template v-if="showButtonText" #default>
-			{{ endCallLabel }}
-		</template>
-	</NcButton>
-	<NcButton
-		v-else-if="showLeaveCallButton && (!canEndForAll || isVoiceRoom) && !isBreakoutRoom"
-		:aria-label="leaveCallLabel"
-		class="leave-call"
-		:variant="isScreensharing ? 'tertiary' : 'error'"
-		:disabled="loading"
-		@click="leaveCall(false)">
-		<template #icon>
-			<NcLoadingIcon v-if="loading" :size="20" />
-			<IconPhoneHangupOutline v-else :size="20" />
-		</template>
-		<template v-if="showButtonText" #default>
-			{{ leaveCallLabel }}
-		</template>
-	</NcButton>
-	<NcActions
-		v-else-if="showLeaveCallButton && (canEndForAll || isBreakoutRoom)"
-		class="leave-call leave-call-actions--split"
-		:disabled="loading"
-		:forceName="showButtonText"
-		placement="top-end"
-		:aria-label="leaveCallActionsLabel"
-		:inline="1"
-		:variant="leaveCallButtonVariant">
-		<template #icon>
-			<IconChevronUp :size="20" />
-		</template>
-		<NcActionButton
-			v-if="isBreakoutRoom"
-			:aria-label="backToMainRoomLabel"
-			@click="switchToParentRoom">
-			<template #icon>
-				<IconArrowLeft class="bidirectional-icon" :size="20" />
-			</template>
-			<template v-if="showButtonText" #default>
-				{{ backToMainRoomLabel }}
-			</template>
-		</NcActionButton>
-		<NcActionButton
-			class="leave-call-button--split"
-			:aria-label="leaveCallLabel"
-			@click="leaveCall(false)">
-			<template #icon>
-				<NcLoadingIcon v-if="loading" :size="20" />
-				<IconPhoneHangupOutline v-else :size="20" />
-			</template>
-			<template v-if="showButtonText || isBreakoutRoom" #default>
-				{{ leaveCallLabel }}
-			</template>
-		</NcActionButton>
-		<NcActionButton v-if="canEndForAll && !isVoiceRoom" @click="leaveCall(true)">
-			<template #icon>
-				<IconPhoneOffOutline :size="20" />
-			</template>
-			{{ t('spreed', 'End call for everyone') }}
-		</NcActionButton>
-	</NcActions>
-</template>
-
 <script>
 import axios from '@nextcloud/axios'
 import { showError } from '@nextcloud/dialogs'
@@ -104,25 +10,16 @@ import { emit } from '@nextcloud/event-bus'
 import { t } from '@nextcloud/l10n'
 import { generateOcsUrl } from '@nextcloud/router'
 import { useIsMobile } from '@nextcloud/vue/composables/useIsMobile'
-import NcActionButton from '@nextcloud/vue/components/NcActionButton'
-import NcActions from '@nextcloud/vue/components/NcActions'
 import NcButton from '@nextcloud/vue/components/NcButton'
 import NcLoadingIcon from '@nextcloud/vue/components/NcLoadingIcon'
-import IconArrowLeft from 'vue-material-design-icons/ArrowLeft.vue'
-import IconChevronUp from 'vue-material-design-icons/ChevronUp.vue'
 import IconPhone from 'vue-material-design-icons/Phone.vue' // Filled used for non-silent calls
 import IconPhoneDialOutline from 'vue-material-design-icons/PhoneDialOutline.vue'
-import IconPhoneHangupOutline from 'vue-material-design-icons/PhoneHangupOutline.vue'
-import IconPhoneOffOutline from 'vue-material-design-icons/PhoneOffOutline.vue'
 import IconPhoneOutline from 'vue-material-design-icons/PhoneOutline.vue'
 import { useGetToken } from '../../composables/useGetToken.ts'
 import { useIsInCall } from '../../composables/useIsInCall.js'
 import { useJoinCall } from '../../composables/useJoinCall.ts'
-import { CALL, CONVERSATION, PARTICIPANT } from '../../constants.ts'
+import { CALL, CONVERSATION } from '../../constants.ts'
 import { getTalkConfig, hasTalkFeature } from '../../services/CapabilitiesManager.ts'
-import { EventBus } from '../../services/EventBus.ts'
-import { useActorStore } from '../../stores/actor.ts'
-import { useBreakoutRoomsStore } from '../../stores/breakoutRooms.ts'
 import { useCallViewStore } from '../../stores/callView.ts'
 import { useSettingsStore } from '../../stores/settings.ts'
 import { useSoundsStore } from '../../stores/sounds.js'
@@ -136,16 +33,10 @@ export default {
 	name: 'CallButton',
 
 	components: {
-		NcActions,
-		NcActionButton,
 		NcButton,
 		// Icons
-		IconArrowLeft,
-		IconChevronUp,
 		IconPhone,
 		IconPhoneDialOutline,
-		IconPhoneHangupOutline,
-		IconPhoneOffOutline,
 		IconPhoneOutline,
 		NcLoadingIcon,
 	},
@@ -184,11 +75,6 @@ export default {
 			default: false,
 		},
 
-		isScreensharing: {
-			type: Boolean,
-			default: false,
-		},
-
 		/**
 		 * Whether to use text on button (e.g. at sidebar)
 		 */
@@ -209,11 +95,9 @@ export default {
 	setup() {
 		const { joinCall } = useJoinCall()
 		return {
-			actorStore: useActorStore(),
 			tokenStore: useTokenStore(),
 			token: useGetToken(),
 			isInCall: useIsInCall(),
-			breakoutRoomsStore: useBreakoutRoomsStore(),
 			callViewStore: useCallViewStore(),
 			talkHashStore: useTalkHashStore(),
 			settingsStore: useSettingsStore(),
@@ -257,17 +141,6 @@ export default {
 			return this.settingsStore.showMediaSettings
 		},
 
-		participantType() {
-			return this.conversation.participantType
-		},
-
-		canEndForAll() {
-			return (this.participantType === PARTICIPANT.TYPE.OWNER
-				|| this.participantType === PARTICIPANT.TYPE.MODERATOR
-				|| this.participantType === PARTICIPANT.TYPE.GUEST_MODERATOR)
-			&& !this.isBreakoutRoom
-		},
-
 		hasCall() {
 			return this.conversation.hasCall
 		},
@@ -283,18 +156,6 @@ export default {
 				|| blockCalls
 		},
 
-		leaveCallLabel() {
-			return t('spreed', 'Leave call')
-		},
-
-		backToMainRoomLabel() {
-			return t('spreed', 'Back to main room')
-		},
-
-		leaveCallActionsLabel() {
-			return t('spreed', 'More actions')
-		},
-
 		startCallLabel() {
 			if (this.hasCall && !this.isInLobby) {
 				return t('spreed', 'Join call')
@@ -305,10 +166,6 @@ export default {
 			}
 
 			return this.silentCall ? t('spreed', 'Start call silently') : t('spreed', 'Start call')
-		},
-
-		endCallLabel() {
-			return t('spreed', 'End call')
 		},
 
 		startCallTitle() {
@@ -339,15 +196,6 @@ export default {
 				&& !this.isInCall
 		},
 
-		showLeaveCallButton() {
-			return this.conversation.readOnly === CONVERSATION.STATE.READ_WRITE
-				&& this.isInCall
-		},
-
-		isBreakoutRoom() {
-			return this.conversation.objectType === CONVERSATION.OBJECT_TYPE.BREAKOUT_ROOM
-		},
-
 		isPhoneRoom() {
 			return isConversationPhoneRoom(this.conversation)
 		},
@@ -358,17 +206,6 @@ export default {
 
 		isJoiningCall() {
 			return this.$store.getters.isJoiningCall(this.token)
-		},
-
-		leaveCallButtonVariant() {
-			if (this.isScreensharing) {
-				return 'tertiary'
-			}
-			return this.isBreakoutRoom ? 'primary' : 'error'
-		},
-
-		isVoiceRoom() {
-			return Boolean(this.conversation.attributes & CONVERSATION.ATTRIBUTE.VOICE_ROOM)
 		},
 	},
 
@@ -388,37 +225,6 @@ export default {
 				silent: this.hasCall ? true : this.silentCall,
 				recordingConsent: this.recordingConsentGiven,
 				shouldStartRecording: this.isRecordingFromStart,
-			})
-			this.loading = false
-		},
-
-		async leaveCall(endMeetingForAll = false) {
-			if (endMeetingForAll) {
-				console.info('End meeting for everyone')
-			} else {
-				console.info('Leaving call')
-			}
-
-			if (this.isVoiceRoom) {
-				this.$router.push({ name: 'root' })
-				// Call ending is handled in App.vue
-				return
-			}
-
-			// Remove selected participant
-			this.callViewStore.setSelectedVideoPeerId(null)
-			this.loading = true
-
-			// Open navigation
-			if (!this.isMobile) {
-				emit('toggle-navigation', {
-					open: true,
-				})
-			}
-			await this.$store.dispatch('leaveCall', {
-				token: this.token,
-				participantIdentifier: this.actorStore.participantIdentifier,
-				all: endMeetingForAll,
 			})
 			this.loading = false
 		},
@@ -470,15 +276,30 @@ export default {
 				this.loading = false
 			}
 		},
-
-		async switchToParentRoom() {
-			EventBus.emit('switch-to-conversation', {
-				token: this.breakoutRoomsStore.getParentRoomToken(this.token),
-			})
-		},
 	},
 }
 </script>
+
+<template>
+	<NcButton
+		v-if="showStartCallButton"
+		:title="startCallTitle"
+		:aria-label="startCallLabel"
+		:disabled="startCallButtonDisabled || loading || isJoiningCall"
+		class="join-call"
+		:variant="hasCall ? 'success' : 'primary'"
+		@click="handleClick">
+		<template #icon>
+			<NcLoadingIcon v-if="isJoiningCall || loading" :size="20" />
+			<IconPhoneDialOutline v-else-if="isPhoneRoom" :size="20" />
+			<IconPhoneOutline v-else-if="silentCall" :size="20" />
+			<IconPhone v-else :size="20" />
+		</template>
+		<template v-if="showButtonText" #default>
+			{{ startCallLabel }}
+		</template>
+	</NcButton>
+</template>
 
 <style lang="scss" scoped>
 .join-call.button-vue--success {
@@ -503,32 +324,4 @@ export default {
 		background-color: var(--join-call-border-color);
 	}
 }
-
-.leave-call.button-vue--error,
-.leave-call :deep(.button-vue--error) {
-	// Overwrite default button colors for leaving call
-	background-color: #FF3333 !important; // Nextcloud 31 --color-error
-	color: var(--color-primary-text) !important;
-
-	&:hover:not(:disabled) {
-		background-color: var(--color-error-hover) !important;
-	}
-}
-
-.leave-call-actions--split {
-	gap: 1px !important;
-}
-
-.leave-call-actions--split :deep(.action-item--single) {
-	border-start-end-radius: 2px;
-	border-end-end-radius: 2px;
-}
-
-.leave-call-actions--split :deep(.action-item__menutoggle) {
-	--button-size: var(--clickable-area-small);
-	height: var(--default-clickable-area);
-	border-start-start-radius: 2px;
-	border-end-start-radius: 2px;
-}
-
 </style>
